Add the configuration files required to generate web site (HTML) output
This commit adds the files that Gamemaker created when I asked it to generate HTML output.

I also changed the label "Shots Remaining" to "Shots remaining" just to verify that my output was showing the latest code.

There are some changes to the main game file (.yyp) that seem to be because the files are being edited in both Windows and Mac OS.